(wip) generate symlink from unsupported files to their dependents instead of generating links
