+++ src/links/node-modules-linker.js
@@ -11,7 +11,8 @@ import ComponentMap from '../consumer/bit-map/component-map';
 import logger from '../logger/logger';
 import { pathRelativeLinux, first } from '../utils';
 import Consumer from '../consumer/consumer';
-import { getLinkContent, getIndexFileName, writeComponentsDependenciesLinks } from './link-generator';
+import { getIndexFileName, writeComponentsDependenciesLinks } from './link-generator';
+import getLinkContent from './link-content';
 import type { PathOsBased } from '../utils/path';
 import GeneralError from '../error/general-error';
 import { Dependency } from '../consumer/component/dependencies';
